Move decl of Outputs to Output.h
